ENH: Implement UpdateKeyFrame button
TODO: right now modifying the camera at the keyframe is not enough
since it does not update the splines automatically.
Right now we reset the same camera to the keyframe and ignore if
the pointer is the same, but the best would be to have an event
listening to the camera modifications.