ENH: set camera clipping range to distance x 6
By default ResetClippingRange set the maximum
clipping to the distance (position to focalpoint)
times two. This is too short in the case of the
flight interactor style since we are moving the
focal point along.

This is because the MRMLCamera does not have
access to the vtkRenderer, which is the one that
can reset the clipping range of the camera.