ENH: Call travelToTime value directly if slider already set to value
When using the GoToKeyFrame button, we call 'timeSlider->SetValue(fNbr)'
to the associated frame of the time, in order to call the observer
method onTimeSliderValueChanged(). If the value to set is the current
one, the method is not called and the camera is not placed to the
keyframe.

travelToTime() can now be called if the value is the current slider
value.