BUG: Use icons in current module directory
I was using Icons that exist in Slicer directory and was
accessing them using the qrc in Slicer sources. This was
an issue since Slicer sources are located in the superbuild
directory and therefore were dependent on my system path.

I added the icons into the module. In the future, if the
module is integrated within Slicer, we could remove them
and use the Slicer core ones.